requirements.txt
Flask
Flask-Babel
google-generativeai
python-dotenv
